Created CardWithImage component from Card function to be able to reuse it legacy.mdx Created legacy docs redirection landing page. Added legacy.mdx to build sidebars.js
